+++ stories/ValidateFloatInput.js
@@ -4,15 +4,44 @@ import React from 'react';
 import { storiesOf, action } from '@kadira/storybook';
 import ValidateFloatInput from '../src/components/ValidateFloatInput';
 
+class Text extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+    this.handleChange = this.handleChange.bind(this);
+  }
+  handleChange(value) {
+    this.setState({
+      foo: value,
+    });
+  }
+  render() {
+    return (
+      <ValidateFloatInput
+        value="0.00"
+        onChange={this.handleChange}
+      />
+    );
+  }
+}
+
 storiesOf('ValidateFloatInput', module)
   .addWithInfo('default', () => (
     <ValidateFloatInput />
   ))
   .addWithInfo('events', () => (
-    <ValidateFloatInput
-      onChange={action('onChange')}
-      onBlur={action('onBlur')}
-    />
+    <div>
+      <ValidateFloatInput
+        onChange={action('onChange')}
+        onBlur={action('onBlur')}
+      />
+      <ValidateFloatInput
+        value="0.00"
+        onChange={action('onChange')}
+        onBlur={action('onBlur')}
+      />
+      <Text />
+    </div>
   ))
   .addWithInfo('show error', () => {
     let vfiRef = null;